:bulb: 동네선택 화면 cell 분리
